Change: Extract a NVDApi from CVEApi
A common base class can be used for the CVE and CPE api.
